Refactor: Split calendar.js from index.js
